[S05P13C107-90] [BE-kskyu610] feat: OotdService 인스턴스 필터링된 ootd 리스트 조회 기능 구현

+++ backend/mabcci/src/main/java/com/mabcci/domain/ootd/dto/OotdListResponse.java
@@ -10,12 +10,12 @@ public final class OotdListResponse {
     private List<OotdResponse> ootdResponses;
 
     @PositiveOrZero
-    private Long totalPages;
+    private int totalPages;
 
     private OotdListResponse() {
     }
 
-    public OotdListResponse(final List<OotdResponse> ootdResponses, final Long totalPages) {
+    public OotdListResponse(final List<OotdResponse> ootdResponses, final int totalPages) {
         this.ootdResponses = ootdResponses;
         this.totalPages = totalPages;
     }
@@ -24,7 +24,7 @@ public final class OotdListResponse {
         return ootdResponses;
     }
 
-    public final Long getTotalPages() {
+    public final int getTotalPages() {
         return totalPages;
     }
 }
